+++ src/main/java/com/example/demo/models/DTO/CommentsDTO.java
@@ -4,7 +4,7 @@ public class CommentsDTO {
 
     private String description;
 
-    private int postId;
+    private long postId;
 
     public CommentsDTO() {
     }
@@ -22,11 +22,11 @@ public class CommentsDTO {
     }
 
 
-    public int getPostId() {
+    public long getPostId() {
         return postId;
     }
 
-    public void setPostId(int postId) {
+    public void setPostId(long postId) {
         this.postId = postId;
     }
 }
