+++ src/main/java/com/example/demo/repositories/UserRepository.java
@@ -16,7 +16,7 @@ public interface UserRepository extends JpaRepository<User, Serializable> {
     User getUserByUsername(String username);
 
     @Query("SELECT u FROM User u WHERE u.id =?1")
-    User getById(int id);
+    User getById(long id);
 
     @Query("SELECT u FROM User u WHERE u.username =?1")
     User getUserByEmail(String email);
@@ -24,9 +24,9 @@ public interface UserRepository extends JpaRepository<User, Serializable> {
     @Transactional
     @Query("UPDATE User u set u.enabled = 0 where u.id = ?1")
     @Modifying
-    void deleteUser(int id);
+    void deleteUser(long id);
 
 
     @Query(value = "select first_name,last_name,username,email,picture,age from users_friends join users u on users_friends.friend_id = u.id where user_id = ?1", nativeQuery = true)
-    List<User>getUserFriendsByUSerId(int userId);
+    List<User>getUserFriendsByUSerId(long userId);
 }
