+++ src/main/java/com/example/demo/services/ImageService.java
@@ -3,7 +3,7 @@ package com.example.demo.services;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface ImageService {
-    void savePostPhoto(int postId, MultipartFile file);
+    void savePostPhoto(long postId, MultipartFile file);
 
-    void saveUserPhoto(int userId, MultipartFile file);
+    void saveUserPhoto(long userId, MultipartFile file);
 }
